user details

+++ services/users/handlers/usersHandler.js
@@ -63,7 +63,7 @@ exports.getAll = async (req, res) => {
 }
 exports.getOne = async (req, res) => {
     try {
-        const user = await User.findById(req.params.id)
+        const user = await User.findById(req.params.id).select('fullName email picture')
         res.status(200).json({ status: 'success', data: { user } })
     } catch (err) {
         console.log(err);
@@ -72,10 +72,10 @@ exports.getOne = async (req, res) => {
 }
 exports.update = async (req, res) => {
     try {
-        if (req.file) {
-            req.body.picture = req.file.filename
-        }
         const user = await User.findById(req.params.id);
+        if (!user) {
+            return res.status(404).send('User not found');
+        }
         if (user.picture !== req.body.picture && user.picture !== 'default.png') {
             await unlink(user.picture)
         }
